Removes the duplication of a config and a blackboard
This moves the logic to auto import into the blackboard. This may not be the smartest long term move but cleans up the logic a bit for now.

Fixes #1
